+++ packages/adhere-util-broadcastchannel-io/es/Response.d.ts
@@ -1,4 +1,4 @@
-import type { ResponseOptions, StateCode } from './types';
+import type { Headers, ResponseOptions, StateCode } from './types';
 /**
  * 响应对象类
  * @class Response
@@ -8,7 +8,7 @@ declare class Response {
     /** 请求ID */
     private requestId;
     /** 响应头 */
-    private readonly headers;
+    private headers;
     /** 状态码 */
     private statusCode;
     /** 状态消息 */
@@ -30,7 +30,7 @@ declare class Response {
      * 获取响应头
      * @returns 响应头的副本
      */
-    getHeaders(): Record<string, string>;
+    getHeaders(): Headers;
     /**
      * 获取响应体
      * @returns 响应体数据