+++ packages/adhere-util-broadcastchannel-io/lib/Constant.d.ts
@@ -7,6 +7,16 @@ export declare const OK_MESSAGE: string;
  */
 export declare const ERROR_MESSAGE: string;
 /**
+ * 心跳服务的action名称
+ */
+export declare const KEEP_ALIVE_SERVICE_NAME = "_keep-alive";
+/**
  * CHANNEL_NAME
  */
 export declare const CHANNEL_NAME = "adhere-util-broadcastchannel-io";
+export declare const STATUS_CODE_INIT = 0;
+export declare const STATUS_CODE_OK = 200;
+export declare const STATUS_CODE_ERROR = 500;
+export declare const STATUS_CODE_TIME_OUT = 408;
+export declare const STATUS_CODE_NOT_FOUND = 404;
+export declare const STATUS_CODE_NOT_ACCEPTABLE = 406;