+++ packages/adhere-util-broadcastchannel-io/src/Constant.ts
@@ -9,6 +9,18 @@ export const OK_MESSAGE: string = 'ok';
 export const ERROR_MESSAGE: string = 'error';
 
 /**
+ * 心跳服务的action名称
+ */
+export const KEEP_ALIVE_SERVICE_NAME = '_keep-alive';
+
+/**
  * CHANNEL_NAME
  */
-export const CHANNEL_NAME = "adhere-util-broadcastchannel-io";
+export const CHANNEL_NAME = 'adhere-util-broadcastchannel-io';
+
+export const STATUS_CODE_INIT = 0;
+export const STATUS_CODE_OK = 200;
+export const STATUS_CODE_ERROR = 500;
+export const STATUS_CODE_TIME_OUT = 408;
+export const STATUS_CODE_NOT_FOUND = 404;
+export const STATUS_CODE_NOT_ACCEPTABLE = 406;