+++ packages/adhere-util-iframeio/es/server/index.d.ts
@@ -38,10 +38,22 @@ declare class Server {
      */
     private service;
     /**
+     * keepAlive
+     * @description 心跳的接口
+     * @private
+     */
+    private keepAlive;
+    /**
      * 启动服务
+     * @param {{
+     *   // 是否开启KeepAlive
+     *   startKeepAlive?: boolean;
+     * }}
      * @returns Promise<void>
      */
-    start(): Promise<void>;
+    start({ startKeepAlive }: {
+        startKeepAlive?: boolean;
+    }): Promise<void>;
     /**
      * 关闭服务
      * @returns Promise<void>