+++ packages/adhere-util-iframeio/lib/types.d.ts
@@ -2,9 +2,13 @@ import type Context from './Context';
 import Request from './Request';
 import Response from './Response';
 /**
+ * Headers
+ */
+export type Headers = Record<string, string>;
+/**
  * 状态码类型定义
  */
-export type StateCode = 0 | 200 | 404 | 500;
+export type StateCode = 0 | 200 | 500 | 408 | 404 | 406;
 /**
  * 上下文选项接口
  */
@@ -28,7 +32,9 @@ export interface SendOptions {
     /** 发送的数据 */
     data?: any;
     /** 请求头 */
-    headers?: Record<string, string>;
+    headers?: Headers;
+    /** 超时时间 */
+    timeOut?: number;
 }
 /**
  * 请求选项接口
@@ -37,7 +43,7 @@ export interface RequestOptions {
     /** 请求路径 */
     pathname: string;
     /** 请求头 */
-    headers?: Record<string, string>;
+    headers?: Headers;
     /** 状态码 */
     statusCode?: StateCode;
     /** 状态消息 */
@@ -54,7 +60,7 @@ export interface ResponseOptions {
     /** 请求ID */
     requestId: string;
     /** 响应头 */
-    headers: Record<string, string>;
+    headers: Headers;
     /** 状态码 */
     statusCode: StateCode;
     /** 状态消息 */
@@ -75,7 +81,7 @@ export interface MessageEventData {
     /** 路径名 */
     pathname?: string;
     /** 请求头 */
-    headers?: Record<string, string>;
+    headers?: Headers;
     /** 状态码 */
     statusCode?: StateCode;
     /** 状态消息 */
@@ -113,6 +119,7 @@ export interface UseServerConfig {
     controllers: RouteController[];
     /** 启动后的回调函数 */
     startAfterCB?: () => void;
+    startKeepAlive?: boolean;
 }
 /**
  * Fetch 配置接口